requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
fastapi==0.115.9
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
Mako==1.3.9
MarkupSafe==3.0.2
openai==0.27.8
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.22.0
pytest==8.3.5